ml_api/requirements.txt
flask
flask-cors
torch
torchvision
timm
pillow
gunicorn